requirements.txt
tqdm==4.67.1
openpyxl==3.1.5
opencv-python==4.11.0.86
numpy==2.1.3
colorama==0.4.6
doxapy==0.9.2
pandas==2.2.3
numba==0.61.0
pillow==11.1.0